+++ sa_modelcrud/crud/base.py
@@ -1,4 +1,4 @@
-from abc import ABC, abstractmethod
+from abc import ABC, abstractproperty
 from datetime import datetime
 from typing import (
     Any,
@@ -26,17 +26,8 @@ UpdateSchemaType = TypeVar("UpdateSchemaType", bound=BaseModel)
 
 
 class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType], ABC):
-    # def __init__(self, model: Type[ModelType]) -> None:
-    #     """CRUD object with default methods to Create, Read, Update, & Delete
-
-    #     Args:
-    #         model (Type[ModelType]): A SQLAlchemy model class
-    #     """
-    #     self.model = model
-
-    @property
-    @abstractmethod
-    def model(self) -> type[ModelType]:
+    @abstractproperty
+    def model(self) -> Type[ModelType]:
         ...
 
     async def get(self, db: AsyncSession, uid: UUID) -> Optional[ModelType]: